requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn==1.3.0
hdbscan==0.8.33
matplotlib>=3.7.5
seaborn==0.12.2
# Embedding and LLM
openai>=1.6.1
google-generativeai>=0.4.1
# Updated LangChain packages
langchain>=0.3.0
langchain-core>=0.3.66
langsmith>=0.1.125
langchain-openai>=0.1.0
langchain-google-genai>=1.0.1
chroma-hnswlib==0.7.3
chromadb>=0.5.0
# RAG and evaluation
trulens-eval>=2.2.1

# Web dashboard
streamlit>=1.35,<2.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
# Testing
pytest==7.4.0